+++ lib/jms/logger.rb
@@ -1,4 +1,0 @@
-# Add trace method to Ruby Logger class
-class Logger
-  alias :trace :debug
-end